Factor out conditional-form-fields.tsx.
